refactored to separate out into common and class files
