+++ src/Student.py
@@ -1,7 +1,5 @@
 from random import randint
-
-PROCRASTINATION_RANGE = (0, 10)
-CHEAT_RANGE = (0, 10)
+from common import *
 
 class Student:
     def __init__(self, id, cheat_level, procrastinate_level, dueDate):
@@ -26,7 +24,7 @@ class Student:
         self.workPerDay = 1 / (self.finishDay - self.startDay + 1)  # get the work done evenly during each day bewteen due date and finish date
 
         self.friends = [] # TODO LATER
-        
+
     def getId(self):
         return self.id
     
@@ -64,7 +62,7 @@ class Student:
             self.finished = True
         
         # cheat if need be
-        self.potentiallyCheat(today, report)
+        # self.potentiallyCheat(today, report)
 
     def potentiallyCheat(self, today, report=False):
         if self.progress < 1 and self.cheat_level <= 3 and today >= self.startDay:
@@ -92,36 +90,7 @@ class Student:
             if report:
                 print('Student ' + str(self.id) + ' received homework from ' + str(friendId))
 
-class Class:
-    def __init__(self, student_count, dueDate):
-        self.dueDate = dueDate # considered to be due at 12:01 AM on the dueDate. For example, if it is due on day 10, they have days 0-9 to work on the assignment
-        self.today = 0
-        self.students = []
 
-        for i in range(student_count):
-            cheat = randint(*CHEAT_RANGE)
-            procrastinate = randint(*PROCRASTINATION_RANGE)
-            self.students.append(Student(i, cheat, procrastinate, dueDate))
-
-    def useDay(self, report=False):
-        if report:
-            print('During day: ' + str(self.today))
-        for student in self.students:
-            student.useDay(self.today, report)
-        if report:
-            self.report()
-        self.today += 1
-
-    
-    def report(self):
-        print("\nAt the end of day: " + str(self.today))
-        for student in self.students:
-            if student.finished:
-                print(f'Student {student.id} has finished')
-            else:
-                print(f'Student {student.id} has progress {student.progress}')
-        print()
-
