+++ src/common.py
@@ -1,0 +1,2 @@
+PROCRASTINATION_RANGE = (0, 10)
+CHEAT_RANGE = (0, 10)
