+++ src/main.py
@@ -1,5 +1,5 @@
-from Student import Student, Class
-from random import randint
+from Student import Student
+from Class import Class
 
 NUM_STUDENTS = 5
 NUM_DAYS = 10
